+++ .gitignore
@@ -384,7 +384,7 @@ gradle-app.setting
 # Avoid ignoring Gradle wrapper jar file (.jar files are usually ignored)
 !gradle-wrapper.jar
 
-# Avoid ignore Gradle wrappper properties
+# Avoid ignore Gradle wrappper property
 !gradle-wrapper.properties
 
 # Cache of project
