+++ src/main/kotlin/cc/polyfrost/polyui/event/Event.kt
@@ -1,0 +1,4 @@
+package cc.polyfrost.polyui.event
+
+// wow, much empty
+interface Event
