+++ src/main/kotlin/cc/polyfrost/polyui/event/FocusedEvent.kt
@@ -1,0 +1,6 @@
+package cc.polyfrost.polyui.event
+
+open class FocusedEvent : Event {
+    data class KeyPressed(val key: Int) : FocusedEvent()
+    data class KeyReleased(val key: Int) : FocusedEvent()
+}
