--- src/main/kotlin/cc/polyfrost/polyui/events/Event.kt
@@ -1,4 +1,0 @@
-package cc.polyfrost.polyui.events
-
-// wow, much empty
-interface Event
