--- src/main/kotlin/cc/polyfrost/polyui/events/FocusedEvent.kt
@@ -1,6 +1,0 @@
-package cc.polyfrost.polyui.events
-
-open class FocusedEvent : Event {
-    data class KeyPressed(val key: Int) : FocusedEvent()
-    data class KeyReleased(val key: Int) : FocusedEvent()
-}
