--- src/main/kotlin/cc/polyfrost/polyui/properties/impl/ImageBlockProperties.kt
@@ -1,9 +1,0 @@
-package cc.polyfrost.polyui.properties.impl
-
-import cc.polyfrost.polyui.color.Color
-import cc.polyfrost.polyui.properties.Properties
-
-open class ImageBlockProperties : Properties() {
-    override val color: Color = Color.WHITE
-    override val padding: Float = 0F
-}
