+++ src/main/kotlin/cc/polyfrost/polyui/property/impl/ImageBlockProperties.kt
@@ -1,0 +1,9 @@
+package cc.polyfrost.polyui.property.impl
+
+import cc.polyfrost.polyui.color.Color
+import cc.polyfrost.polyui.property.Properties
+
+open class ImageBlockProperties : Properties() {
+    override val color: Color = Color.WHITE
+    override val padding: Float = 0F
+}
