+++ src/main/kotlin/cc/polyfrost/polyui/renderer/Window.kt
@@ -7,7 +7,7 @@ import cc.polyfrost.polyui.PolyUI
  * As a rendering implementation, you must:
  *  - implement all the methods
  *  - create callbacks for the event methods in PolyUI e.g. [PolyUI.onResize]
- *  - create callbacks for all the event-related methods in [cc.polyfrost.polyui.events.EventManager]
+ *  - create callbacks for all the event-related methods in [cc.polyfrost.polyui.event.EventManager]
  *  - call [open] to start the rendering loop; this can be blocking or non-blocking. Please note that after open is called, the rendering implementation will be created. This means that in a thread-based system such as LWJGL's OpenGL, you **must** ensure that it is fully setup before exiting `init {}`.
  */
 abstract class Window(var title: String, var width: Int, var height: Int) {
@@ -17,7 +17,7 @@ abstract class Window(var title: String, var width: Int, var height: Int) {
     protected abstract fun renameWindow(title: String)
     abstract fun fullscreen()
 
-    /** Create the callbacks for window events.
+    /** Create the callbacks for window event.
      * @see Window */
     abstract fun createCallbacks()
