+++ src/main/kotlin/cc/polyfrost/polyui/renderer/data/Framebuffer.kt
@@ -1,6 +1,6 @@
 package cc.polyfrost.polyui.renderer.data
 
-import cc.polyfrost.polyui.properties.Settings
+import cc.polyfrost.polyui.property.Settings
 
 data class Framebuffer(val width: Float, val height: Float, val type: Settings.BufferType) {
     enum class Mode {
