+++ src/main/kotlin/cc/polyfrost/polyui/unit/Position.kt
@@ -1,0 +1,6 @@
+package cc.polyfrost.polyui.unit
+
+enum class Position {
+    Above, Below, Left, Right, Center,
+    Top, Bottom;
+}
