--- src/main/kotlin/cc/polyfrost/polyui/units/Position.kt
@@ -1,6 +1,0 @@
-package cc.polyfrost.polyui.units
-
-enum class Position {
-    Above, Below, Left, Right, Center,
-    Top, Bottom;
-}
